Make the Jit Graph use dynamic arrays
This is a large, but not very exciting change; it implies that the
jit graph is now malleable for places that don't have (or are not
interested in) the jit graph builder; specifically the jit graph
can now add and find labels for itself, which I need for the
expression compiler.

It also leaks tiny amounts of memory, which will be fixed in due
course.